backend/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.0.0
flask-cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
